requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers